refactor: extract blur and linking helpers
